+++ NEWS.md
@@ -3,6 +3,8 @@ Petersen 2024.01.01
 -------------------
 
 * Some typos fixed in documentation. No new functionality added.
+* New example of reverse capture-recapture geographically stratified study on
+lower Fraser River coho
 
 
 Petersen 2023.12.01